Save speech message working
After the recording stops, the frontend asks Database class to store the message into the actual db, for it to work the Database class had to receive a good update. Interview scene is now being checked when loaded to see if the user is the actual owner of the interview.